requirements.txt
yt-dlp==2025.7.21
openai-whisper==20231117
openai>=1.0.0
python-dotenv